requirements.txt
# Core dependencies
anthropic>=0.18.0
tweepy>=4.14.0
python-dotenv>=1.0.0
APScheduler>=3.10.4

# Database
# sqlite3 is part of Python's standard library, no need to install

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2

# Logging and monitoring
colorlog>=6.8.0

# Data validation
pydantic>=2.5.0

# Optional: For enhanced content analysis
textstat>=0.7.3
